ozon.models.Catalog (which contains SearchResultsV2 and Category) is added Fix value of numeric data type for some fields

+++ src/main/scala/marketplace/models/ozon/Delivery.scala
@@ -13,7 +13,7 @@ import io.circe.Decoder
 
 @derive(loggable)
 @AvroNamespace("ozon.models")
-final case class Delivery(schema: Delivery.Schema, timeDiffDays: Int)
+final case class Delivery(schema: Delivery.Schema, timeDiffDays: Short)
 
 object Delivery {
